# backend/swagger.yml
openapi: 3.0.3
info:
  title: Servnix API
  version: 1.0.0
  description: Account, organization, and appointment endpoints.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /v1/reviews/organization/{id}:
    get:
      tags: [Reviews]
      summary: List public reviews for an approved organization
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Review list }
  /v1/reviews:
    post:
      tags: [Reviews]
      summary: Review an appointment marked completed by the provider
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, rating]
              properties:
                bookingId: { type: string, format: uuid }
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string, maxLength: 1000 }
      responses:
        '201': { description: Review created }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...